+++ amplipi/ctrl.py
@@ -223,7 +223,8 @@ class Api:
       is_streamer=self.is_streamer,
       lms_mode=self.lms_mode,
       version=utils.detect_version(),
-      stream_types_available=amplipi.streams.stream_types_available()
+      stream_types_available=amplipi.streams.stream_types_available(),
+      extra_fields=utils.load_extra_fields()
     )
     for major, minor, ghash, dirty in self._rt.read_versions():
       fw_info = models.FirmwareInfo(version=f'{major}.{minor}', git_hash=f'{ghash:x}', git_dirty=dirty)